Hi i have made changes regarding #2248 and #2397 code review observations for Hide the progress bar with input file popup, User int try parse instead of Convert to int method, and check the file exists or not before remove the file.

+++ RevenuePlanner/Controllers/FinanceController.cs
@@ -4273,7 +4273,7 @@ namespace RevenuePlanner.Controllers
             DataSet ds = new DataSet();
             try
             {
-                
+
                 if (Request != null)
                 {
                     if (Request.Files[0].ContentLength > 0)
@@ -4296,7 +4296,11 @@ namespace RevenuePlanner.Controllers
                                                                                    fileLocation + ";Extended Properties=\"Excel 8.0;HDR=Yes;IMEX=1\"";
                                 IExcelDataReader excelReader = ExcelReaderFactory.CreateBinaryReader(Request.Files[0].InputStream);
                                 ds = excelReader.AsDataSet();
-                                objImprtData = GetXLSData(ds, BudgetDetailId); // Read Data from excel 2003/(.xls) format file to xml
+                                if (ds != null && ds.Tables.Count > 0)
+                                {
+                                    objImprtData = GetXLSData(ds, BudgetDetailId); // Read Data from excel 2003/(.xls) format file to xml
+                                }
+
                                 if (ds == null)
                                 {
                                     return Json(new { msg = "error", error = "Invalid data." }, JsonRequestBehavior.AllowGet);
@@ -4308,7 +4312,11 @@ namespace RevenuePlanner.Controllers
                             }
                             dtColumns = objImprtData.MarketingBudgetColumns;
                             xmlData = objImprtData.XmlData;
-                            System.IO.File.Delete(fileLocation);
+
+                            if (System.IO.File.Exists(fileLocation))
+                            {
+                                System.IO.File.Delete(fileLocation);
+                            }
 
                             if (!string.IsNullOrEmpty(objImprtData.ErrorMsg))
                             {
